build: limit libelf/libbpf dependencies to only tc and ip
All iproute2 utilities link with libutil.a, which includes ELF and BPF
functions that in turn depend on libelf and libbpf. For most utilities
these dependencies are unnecessary, bloating their installation footprint.

Isolate libelf and libbpf related functions in a new library libbpfutil.a,
linked only with tc and ip.

Signed-off-by: Tony Ambardar <Tony.Ambardar@gmail.com>